Move style_assets into separate file
